src/ecr/requirements.txt
coralogix_logger==2.0.5
requests==2.31.0
